split of header and footer code importer
