+++ tools/importer/transformers/index.js
@@ -1,0 +1,21 @@
+import header from './header.js';
+import footer from './footer.js';
+import stickyFooter from './stickyFooter.js';
+
+// eslint-disable-next-line import/prefer-default-export
+export const xfTransformers = [
+  footer,
+  stickyFooter,
+];
+
+export const xfAsyncTransformers = [
+  header,
+];
+
+export const preTransformers = [
+
+];
+
+export const postTransformers = [
+
+];
